+++ packages/amazonq/.changes/next-release/Breaking Change-912fe720-0d1b-40a3-8638-f9a4d2948904.json
@@ -1,0 +1,4 @@
+{
+	"type": "Breaking Change",
+	"description": "Change keybind for focusing chat to ctrl+win+i on Windows, ctrl+cmd+i on macOS and ctrl+meta+i on Linux"
+}
