+++ packages/amazonq/.changes/next-release/Feature-31cd9ad4-1e6b-4f05-b910-11c5059bd2fa.json
@@ -1,0 +1,4 @@
+{
+	"type": "Feature",
+	"description": "Use inline chat to select code and transform it with natural language instructions"
+}
