+++ packages/amazonq/src/inlineChat/output/responseTransformer.ts
@@ -7,6 +7,14 @@ import { getLogger } from 'aws-core-vscode/shared'
 import { decode } from 'he'
 import { InlineTask } from '../controller/inlineTask'
 
+/**
+ * Transforms the response from the INLINE_CHAT GenerateAssistantResponse call.
+ *
+ * @param response - The raw response string from GenerateAssistantResponse.
+ * @param inlineTask - The inline task object containing information about the current task.
+ * @param isWholeResponse - A boolean indicating whether this is a complete response or a partial one.
+ * @returns The decoded response string, or undefined if an error occurs.
+ */
 export function responseTransformer(
     response: string,
     inlineTask: InlineTask,
@@ -28,6 +36,16 @@ export function responseTransformer(
     }
 }
 
+/**
+ * This function is used to handle partial responses in inline tasks. It divides
+ * the selected text into two parts:
+ * 1. The "left" part, which contains the same number of lines as the response.
+ * 2. The "right" part, which contains the remaining lines.
+ *
+ * @param response - The response string from the assistant.
+ * @param inlineTask - The inline task object containing the full selected text.
+ * @returns A tuple with two strings: [leftPart, rightPart].
+ */
 function extractPartialCode(response: string, inlineTask: InlineTask): [string, string] {
     const lineCount = response.split('\n').length
     const splitLines = inlineTask.selectedText.split('\n')
