+++ packages/core/src/amazonq/util/authUtils.ts
@@ -6,6 +6,15 @@
 import { FeatureAuthState } from '../../codewhisperer'
 import { AuthFollowUpType, AuthMessageDataMap } from '../auth/model'
 
+/**
+ * This function evaluates the authentication state of CodeWhisperer features (chat and core)
+ * when the authentication is not valid, and returns the appropriate authentication follow-up type and message.
+ *
+ * @param credentialState - The current authentication state for each CodeWhisperer feature
+ * @returns An object containing:
+ *   - authType: The type of authentication follow-up required (AuthFollowUpType)
+ *   - message: The corresponding message for the determined auth type
+ */
 export function extractAuthFollowUp(credentialState: FeatureAuthState) {
     let authType: AuthFollowUpType = 'full-auth'
     let message = AuthMessageDataMap[authType].message
