+++ packages/core/src/shared/index.ts
@@ -56,3 +56,4 @@ export { CrashMonitoring } from './crashMonitoring'
 export { amazonQDiffScheme } from './constants'
 export * from './featureConfig'
 export * from './icons'
+export * from './utilities/textDocumentUtilities'
